chore(copier): add heavy-backfill divergence markers + cli shim
Builds on the lazy bootstrap (previous commit) with divergence comments
and a cli.py shim so future copier updates can't silently overwrite
project-owned customisations.

- pyproject.toml: license + MIT classifier divergence comments;
  PROJECT-DEPS-START/END + PROJECT-EXTRAS-START/END sentinel zones
  around domain deps + extras.
- packaging/nfpm.yaml: license divergence comment.
- src/image_generation_mcp/cli.py -> 3-line shim that re-exports
  main from _cli_impl.py (project-owned, not in template).
- tests/test_cli.py: imports updated to point at _cli_impl.

A literal copier update from v1.0.0 -> v1.1.0 was destructive:
the license line would silently flip MIT -> LicenseRef-Proprietary
and the argparse CLI would be replaced by the typer scaffold.
